Refactor navbar into smaller reusable components
- Extract NavLink component for reusable navigation links with active state
- Extract Logo component for brand/logo section
- Extract MobileMenuButton component for hamburger menu
- Extract MobileMenu component for mobile navigation
- Extract DesktopNav component for desktop navigation
- Create nav-links.ts for navigation constants
- Move navbar.tsx into components/navbar/ directory for consistency
- Update imports in app/layout.tsx

This refactoring eliminates code duplication, improves maintainability,
and aligns with the component design pattern used throughout the codebase.
