+++ components/navbar/nav-links.ts
@@ -1,0 +1,8 @@
+export const NAV_LINKS = [
+  { href: "/", label: "Home" },
+  { href: "/about", label: "About" },
+  { href: "/what-is-ultimate", label: "What is Ultimate?" },
+  { href: "/news", label: "News" },
+  { href: "/schedule", label: "Schedule" },
+] as const;
+
